Improve HTML structure
Old HTML does not include HTML5 semantic tags and uses a number
of nondescript div elements. This is harder to style in CSS,
is difficult for screenreaders etc and doesn't follow best practice.

New HTML adds semantic HTML5 tags to improve accessability and make
future CSS styling easier.

The new HTML breaks the existing CSS styling so it is also
 necessary to update the CSS stylesheet.