Add initial HTML and CSS for contact form
Split .splash CSS class into .splash-text and .splash-background
to allow the gradient background to be applied to the contact
form without the text going massive as well.

Make first attempts at structure and styling of contact form.
Future adjustments will need to include on hover and active
styling to make the form look nicer and easier to use.