Add correct modal window behaviour
Apply small tweaks to JS, CSS and add extra div
in HTML to get desired modal behaviour. What I was
doing wrong in the previous commit was making the
modal the size of the form itself. The idea is to
make the modal the size of the whole viewport and put
the form on top of that. This way, the JS click on
modal to close form is effectively click anywhere
outside form to close it, as intended.