Add smooth scroll links
Links will now scroll smoothly to destination element rather than
jumping. Done by replacing the href property with an onclick call
to a simple javascript function which scrolls smoothlyto a given
element.

Might not want this behaviour on mobile, if not we will need to
add some js that tests the browser type. Having onclick calls to
a js function on every <a> tag feels a bit clunky. Find out if
there is a way make the js listen for clicks on any element of a
given type/class/id etc.