Apply styling changes from contact-form branch
Contact-form branch contains a number of styling updates that it would be
nice to push to live website. However, the form itself is not ready yet.
This update takes all the styling updates from the contact-form branch but
removes the HTML form so that users don't try and submit to a non-functional
form.