Restructure for Flask app and setup for backend
This commit moves the project from a static site setup to a flask app setup in order to
enable the implementation of a simple backend for the email contact form. Currently the
form renders but the actual email functionality needs to be implemented. Also styling
and validation and so on.