Adds first spam honeypot implementation
This commit adds the first attempt at a spam honeypot implementation on the contact form.
It also adds some basic logging to capture the date, time and email address from which the
spam message was recieved. The logging could probably be improved as at the moment it also
captures a bunch of the wsgi type server info messages as well. Could also write a new log
file every day or week or something.